+++ src/exceptions.ts
@@ -5,3 +5,11 @@ export class DockerConnectionError extends Error {
         Object.setPrototypeOf(this, DockerConnectionError.prototype);
     }
 }
+
+export class DockerNotFoundError extends Error {
+    public constructor(message: string) {
+        super(message);
+
+        Object.setPrototypeOf(this, DockerConnectionError.prototype);
+    }
+}